Move BootROM from Coreplex to Periphery
